+++ src/modules/Elsa.Workflows.Core/Activities/Flowchart/Activities/FlowJoin.cs
@@ -1,5 +1,4 @@
 using System.Runtime.CompilerServices;
-using Elsa.Extensions;
 using Elsa.Workflows.Activities.Flowchart.Contracts;
 using Elsa.Workflows.Activities.Flowchart.Extensions;
 using Elsa.Workflows.Activities.Flowchart.Models;
